This adds logic to better support stackage (stack-to-nix)
When using stackage as a baseline via stack-to-nix, we usually
assume stackage to be the base, and some augemnetation via
extra-dependencies and packages in the `stack.yaml` file.

To achive this we extend the package-set builder by allowing
a list of additional sets of package to speicied. These are then
folded onto the base package set.

As stackage seems to define the `rts` package as well, we force
`rts-1.0` to be null ontop of our hackage database.  As of
today (November 7th, 2018) there is no way of building the rts outside
of GHC; and as such no way to ever get the rts package onto hackage.

Once this package is *on* hackage, we can drop the hack.