Pin the index-state used by ghc dependencies (#264)
* Pin the index-state used by ghc dependencies

Currently when a new hackage index-state is available (every day on
haskell.nix/master branch) the index-state used alex, happy and hscolour
is in the bootPkgs is updates to that latest one.  This results in a
rebuild of GHC (even though the build plans are likely unchanged).

By pinning the index state for these bootPkgs nix will be able to
reuse GHC.  The only down side is that if we need to get new versions
of alex, happy and hscolour into bootPkgs we will likely need to bump
the index-state attributes as well as the version.