Expose plan-nix for alex, happy and hscolour
This exposes the plan nix though:
bootstrap.haskell.packages.alex-project.plan-nix

By building this on hydra you can get better errors and I think it
helps ensure cache stays populated with the dependencies needed to
build the plan.

It is also handy to be able to inspect the nix files that are
generated by plan-to-nix.