Add collectRunComponents. Move checkPhase to .run
In order to make it easier to run tests this change adds a
collectRunComponents function (like collectComponents but
with it returns the `passthru.run` derivation for each component).

This change also disables `checkPhase` for all components to avoid
running the tests twice.  To run a test or benchmark you now have to
use the the `passthru.run` derivation.