Move envDeps and exactDeps into lib and ghc (#360)
This change moves the work done to in `envDeps` and `exactDeps`
from `make-config-files.nix` into the library components and
the `ghc` compiler component (well `passthru` derivations on the
`ghc` component).

This greatly reduces the number of times `runCommand` is
used while nix is working out what to build.

This fixes a performance regression in #323 for builds where
little or no work needs to be done.  While maintaining the
performance improvement in #323 that resulted from reducing the
number of calls to `ghc-pkg` when significant work does need
to be done.