mkStackPkgSet: filter out $locals package
I found that if I added this to my stack.yaml:

    ghc-options:
      $locals: -ddump-to-file -ddump-hi

As specified in https://docs.haskellstack.org/en/stable/yaml_configuration/#ghc-options

Then my build would fail to evaluate with:

    error: The option `packages."\$locals".package.identifier.name' is used but not defined.

So to fix it, we filter out these special package globs in mkStackPkgSet.

This commit also adds a regression test for that configuration option.