Fix linking of dynamic libs with static on Darwin
This piece of code `ln`s all libraries we might want to link with into a
single directory and then hides all previous directories from Cabal.
The problem is that this code only `ln`s dynamic libraries, however we
might want to link with some library that only comes in a static
version.

So, `ln` static `.a` libraries too so that they are found when linking.