Add haskell-nix.tools for common haskell tools
A loosely curated set of tools we can select with:

```
  haskell-nix.tools.cabal
or
  shellFor { tools = ts: with ts; [ cabal ]; }
```

It will be compiled with the same version of ghc used in the
shell (the build ghc in the case of cross compilation).

If want to specify the ghc used to compile. Use:

```
  haskell-nix.tools.ghc883.cabal
or
  shellFor { tools = ts: with ts; [ ghc883.cabal ]; }
```

By default we will get the latest version of the tool.
To specify the version use:

```
  haskell-nix.tools.cabal."3.2.0.0"
or
  shellFor { tools = ts: with ts; [ cabal."3.2.0.0" ]; }
```