Use nix-tools built with the boot ghc by default
While it is possible to use `nix-tools` that version will be built
with the default ghc (currently 8.6.5).  This introduces a
dependency on that compiler in the a number of places.
This can be frustrating if you are using ghc 8.8.3.
Using boot-nix-tools avoids any chancde of need ing
to build or download ghc 8.6.5 (as the boot compiler installed
from a binary distro).