Add niv sources at haskell-nix.sources
I think it would be nice to be able to use niv to manage sources. I
like `niv update` to bump to latest version or switch branches.
It is easy enough to modify the `nix/sources.json` manually if anyone
does not want to learn a new tool and since @michaelpj already added
a niv like `sources` very little changes.  I have exposed the
`sources` as `haskell-nix.sources` so that we can get at it from
wherever we need it.

I've just put the `nixpkgs` pins in `nix/sources.json` for now and
we can add things too it as we need to update them.