Add `lookupSha256` to `stackProject`
The `cabalProject` function has a handy way for sha256 values to
be passed in.  This change adds the same ability to `stackProject`
and `stackProject'`.  The tests also demonstrate how to use
this function both `cabalProject` and `stackProject` to build
the haskell-language-server project providing the sha256 hashes
needed to for this to work with `--option restricted-eval true`.