Remove internal deps on default ghc and stackage
Adds hpack executable to the nix-tools derivation.

Adds a `cabal-hpack` test to make sure `hpack` works with
`cabalProject`.

Issues warnings when compiler-nix-name is not provided.

Reduces the number of calls to `cabalProject` (particularly when
checking materialization), by giving internal tools a per-compiler
attribute.

Adds `p.tool`, `p.tools` and `p.roots` where `p` is a project as an
easy way to match the `cabal-nix-name` of the project.

Uses happy 1.19.12 when building newer ghc versions.

Updates cabal-install 3.2.0.0 to use the source from github that
is compatible with ghc 8.10.1.