fix: remove numa support for armv6l (#808)
There is no support for NUMA on Armv6l (such as in the Raspberry Pi).
Having a dependency on numactl at the compiler-level seems to also
enable numa support for the generated programs as well. Even if it's not
used, dependency on libnuma prevents the generated executables to launch
without an error indicating that libnuma could not be found.