Allow older `ghc` versions in `cabalProject`
Currently the `dummy-ghc-data`, `nix-tools` plan and `cabal-install`
plan must be materialized for the selected `ghc`.  This change
relaxes that restriction and will instead use `nix-tools` and
`cabal-install` built with a fixed `ghc` version (currently `ghc865`)
and will calculate the `dummy-ghc-data` as an IFD.

It will `__trace` a warning of the form:

```
WARNING: No materialized dummy-ghc-data for ghc-8.6.4-x86_64-darwin.
```