Fix issue where ifdLevel with HLS tests
When building a new GHC in haskell.nix it is often desirable to set the
ifdLevel to make sure tests that use eval time do not run.
Unfortunately when ifdLevel=2 eval time work for the Haskell Language
Server tests runs and builds nix-tools for the first time at eval time.

This is fixed by separating the building of the HLS from the
ifdInputs that are needed for the building.