Make sure the a .conf file is created
Had a strange issue where eval of #955 and it might have been caused
by a the nix-tools library seeming to have no `.conf` files.

Since it seemed to go away by itself I think this was a nix GC issue
or something.  Here is some code we could add to the builders to
double check the files are created if it turns out that is not the
case.