LicenseRef-OtherLicense is not `free = false`
The [dependent-map](https://hackage.haskell.org/package/dependent-map)
package for instance is not clear on which free license applies to
which parts of the code, but that does not mean it is not free.

`cabal2nix` maps LicenseRef-OtherLicense to `meta.license = "unknown"`

`cabal-to-nix` and `haskell.nix` map it to a more detailed attribute
set but it includes `free = false` and that causes packages like
`dependent-map` to fail to install when `allowUnfree` has not been set.

This change removes the `free` attribute altogether, which seems more
principled than asserting that the `OtherLicense` cannot be `free`.