Add `copyTextDir` and use on for cabal configure
`cabal configure` for some large haskell projects (such as
`cardano-node`) can take several minutes to run.

When building locally `cleanSourceWith` is used so that only changes to
`cabal.project` and `.cabal` cause `cabal configure` to be run again.

This does not work on hydra though and every commit will cause
`cabal configure` to be rerun during `eval`.  It also means that
local builds cannot reuse the results from hydra.

To fix this `copyTextDir` runs `__readFile` and `writeTextFile` on
the filtered subset of files (cabal.project and .cabal).

Even if the input path is a different store directory the resulting
derivation will be the same if the contents of the filtered subset
of files is the same.