flake: export internal functions for more flexibility
As of d014079 (#953), haskell.nix internals (such as unapplied
./overlays, sources, etc) are not exported in flake.nix. This makes
it harder to override things when using haskell.nix as a flake.

The use-case I have in mind is to pin haskell.nix itself to prevent
unnecessary GHC rebuilds/redownloads and update only
hackage&stackage. This could be done with a following overlay:

    (haskell-nix.overlays {
      sourcesOverride = haskell-nix.sources // {
        inherit hackage stackage;
      };
    }).combined-eval-on-build

However, this is not the case as of the aforementioned commit. Now, to
do this one has to import ./nix/sources.json, reimplementing what's
already done in haskell.nix, and also import ./overlays. With this
commit, previous snippet turns into the following:

    (haskell-nix.internal.overlaysOverrideable {
      sourcesOverride = haskell-nix.internal.sources // {
        inherit hackage stackage;
      };
    }).combined-eval-on-build

`nix flake check` will now warn us of "unknown flake output
'internal'", but I think it is fine (e.g. nixpkgs has an unknown flake
output as well)