Disable benchmarks and tests for tools
By default haskell.nix cabalProject functions run cabal
configure with --enable-tests and --enable-benchmarks.
This is good when you are working on a project as
it means your tests and benchmarks are in the
package `components`.

The `tools` functions (and `shellFor` `tools` arg)
return `components.exes.${toolName}` and while
it is possible to access the other components
via the `project` and `package` properties of
tool component.

In some cases the benchmarks and tests can
have problematic constraints that cause
problems building other components.

Disabling the tests and benchmarks by default
when building `tools` should reduce the these
issues.